requirements.txt
fastapi
uvicorn[standard]
pytesseract
Pillow
python-multipart
PyMuPDF
easyocr
numpy
llama-index-core
llama-index-readers-file
